+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Stringify: Fix extra blank lines when extracting nested inline tables via `inlineTableStart` ([#281])
 - Patching: Fix extra blank lines after tables emptied ([#281])
+- Patching: Fix dotted-key truncation and whitespace bugs ([#282])
 
 ### Changed
 - Improve stringification performance ([#280])
@@ -386,4 +387,6 @@ This first forked version from [timhall/toml-patch](https://github.com/timhall/t
 [#277]: https://github.com/DecimalTurn/toml-patch/pull/277
 [#278]: https://github.com/DecimalTurn/toml-patch/pull/278
 [#280]: https://github.com/DecimalTurn/toml-patch/pull/280
+[#281]: https://github.com/DecimalTurn/toml-patch/pull/281
+[#282]: https://github.com/DecimalTurn/toml-patch/pull/282
 [0e66e68]: https://github.com/DecimalTurn/toml-patch/commit/0e66e68cbf42a07bc44445e46c3ea7bea97f95c1